% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{assemble_connections}
\alias{assemble_connections}
\title{Assemble \code{cicero} coaccessibility relations}
\usage{
assemble_connections(cicero_model_list, silent = FALSE)
}
\description{
Assemble \code{cicero} coaccessibility relations
}
\keyword{internal}
