% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{build_gene_GRanges}
\alias{build_gene_GRanges}
\title{Generate a \code{GRanges} object composed of gene annotations}
\usage{
build_gene_GRanges(org = "hg38")
}
\description{
Generate a \code{GRanges} object composed of gene annotations
}
\keyword{internal}
