% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{calculate_sil}
\alias{calculate_sil}
\title{Calculate the optimization function value}
\usage{
calculate_sil(selected.HBCs, links.df)
}
\description{
Calculate the optimization function value
}
\keyword{internal}
