% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{CollapseToLongestTranscript}
\alias{CollapseToLongestTranscript}
\title{Retain the longest transcripts for all exons}
\usage{
CollapseToLongestTranscript(ranges)
}
\description{
Retain the longest transcripts for all exons
}
\keyword{internal}
