% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{compute_add}
\alias{compute_add}
\title{Calculate the submodular optimization value after adding a hybrid bicluster (HBC)}
\usage{
compute_add(query, hit, sim.m, HBC.max.sim)
}
\description{
Calculate the submodular optimization value after adding a hybrid bicluster (HBC)
}
\keyword{internal}
