% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{compute_original_sim}
\alias{compute_original_sim}
\title{Calculate the pairwise similarity between hybrid biclusters (HBCs)}
\usage{
compute_original_sim(HBCs, features = "genes")
}
\description{
Calculate the pairwise similarity between hybrid biclusters (HBCs)
}
\keyword{internal}
