% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{compute_sim}
\alias{compute_sim}
\title{Calculate pairwise similarity between hybrid biclusters (HBCs)}
\usage{
compute_sim(HBCs)
}
\description{
Calculate pairwise similarity between hybrid biclusters (HBCs)
}
\keyword{internal}
