% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{create_rna_atac}
\alias{create_rna_atac}
\title{Get a simulated jointly profiled scRNA-seq and scATAC-seq data in which
several enhancer regulons (eRegulons)
are contained}
\usage{
create_rna_atac(
  obj = NULL,
  ntfs = 5,
  ngenes = 100,
  ncells = 100,
  all.genes = 1000,
  all.enhs = 3000,
  all.cells = 1000,
  org = "hg38",
  atac.assay = "ATAC",
  gene.links = 2,
  distance = 250000,
  url.link = "https://figshare.com/ndownloader/files/38644505"
)
}
\arguments{
\item{obj}{a \code{Seurat} object used as the prototype to generate simulated dataset
other is \code{GRanges} list}

\item{ntfs}{the number of eRegulons (TFs) to include in the simulated dataset}

\item{ngenes}{the average number of genes in an eRegulon}

\item{ncells}{the average number of cells in an eRegulon}

\item{all.genes}{the number of genes in the simulated \code{Seurat} object}

\item{all.enhs}{the number of enhancers in the simulated \code{Seurat} object}

\item{all.cells}{the number of cells in the simulated \code{Seurat} object}

\item{org}{the organism, e.g., hg38}

\item{atac.assay}{the assay to save the scATAC-seq data}

\item{gene.links}{the average number of enhancers linked to each gene in an eRegulon}

\item{distance}{the maximum distance between a gene and its linked enhancers}
}
\value{
returns a list composed of a eRegulon list and a \code{Seurat} object
}
\description{
Get a simulated jointly profiled scRNA-seq and scATAC-seq data in which
several enhancer regulons (eRegulons)
are contained
}
