% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{DistanceToTSS}
\alias{DistanceToTSS}
\title{Calculate the distance for each enhancer to their nearest TSS,
which was borrowed from Signac}
\usage{
DistanceToTSS(peaks, genes, distance = 2e+05, sep = c("-", "-"))
}
\description{
Calculate the distance for each enhancer to their nearest TSS,
which was borrowed from Signac
}
\keyword{internal}
