% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{ensembl_to_symbol}
\alias{ensembl_to_symbol}
\title{Convert ENSEMBL IDs to gene symbols}
\usage{
ensembl_to_symbol(ensembl.ll, org = org)
}
\description{
Convert ENSEMBL IDs to gene symbols
}
\keyword{internal}
