% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{estimate_distance_parameter}
\alias{estimate_distance_parameter}
\title{Estimate distance parameter using \code{cicero}}
\usage{
estimate_distance_parameter(
  cds,
  window = 5e+05,
  maxit = 100,
  s = 0.75,
  sample_num = 100,
  distance_constraint = 250000,
  distance_parameter_convergence = 1e-22,
  max_elements = 200,
  genomic_coords = cicero::human.hg19.genome,
  max_sample_windows = 500
)
}
\description{
Estimate distance parameter using \code{cicero}
}
\keyword{internal}
