% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{exist_seed}
\alias{exist_seed}
\title{Check whether a seed is eligible}
\usage{
exist_seed(
  s,
  HBCs,
  same.terminal = T,
  intra.cutoff = 1,
  inter.cutoff = 0.8,
  peak.cutoff = 0.8
)
}
\description{
Check whether a seed is eligible
}
\keyword{internal}
