% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{expand_cells}
\alias{expand_cells}
\title{Expand HBC by adding cells}
\usage{
expand_cells(
  HBC = NULL,
  m = NULL,
  mm = NULL,
  quantile.cutoff = 4,
  P = NULL,
  Q = NULL,
  G = NULL,
  KL = "min.exp"
)
}
\description{
Expand HBC by adding cells
}
\keyword{internal}
