% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{expand_core}
\alias{expand_core}
\title{Expand the core part of the hybrid bicluster (HBC)}
\usage{
expand_core(
  HBC,
  top.genes.peaks,
  rna.m,
  atac.m,
  G,
  KL = "min.exp",
  Q = NULL,
  P = NULL,
  DELTA = 0.0015,
  closure = FALSE,
  ego.order = 1,
  min.cells = 10
)
}
\description{
Expand the core part of the hybrid bicluster (HBC)
}
\keyword{internal}
