% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{expand_fuzzy}
\alias{expand_fuzzy}
\title{Add other genes or peaks that are consistent with the hybrid bicluster (HBC)}
\usage{
expand_fuzzy(
  HBC,
  G,
  cand.genes,
  cand.peaks,
  rna.m,
  atac.m,
  g.cutoff = 1,
  m,
  mm,
  ego.order = 1,
  closure = F,
  P = NULL,
  Q = NULL,
  KL = "min.exp"
)
}
\description{
Add other genes or peaks that are consistent with the hybrid bicluster (HBC)
}
\keyword{internal}
