% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{filter_nearby_genes}
\alias{filter_nearby_genes}
\title{Retain genes or enhancers which are within a nearby range of at least one enhancer or gene,
respectively}
\usage{
filter_nearby_genes(obj, distance = 5e+05, peak.assay = "ATAC")
}
\description{
Retain genes or enhancers which are within a nearby range of at least one enhancer or gene,
respectively
}
\keyword{internal}
