% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{find_distance_parameter}
\alias{find_distance_parameter}
\title{Find distance parameter}
\usage{
find_distance_parameter(
  dist_mat,
  gene_range,
  maxit,
  null_rho,
  s,
  distance_constraint,
  distance_parameter_convergence
)
}
\description{
Find distance parameter
}
\keyword{internal}
