% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reguome_tools.R
\name{find_TF_gene}
\alias{find_TF_gene}
\title{Discover cell-subpopulation-active TF-target pairs}
\usage{
find_TF_gene(G.list, bound.TFs, binding.CREs)
}
\description{
Discover cell-subpopulation-active TF-target pairs
}
\keyword{internal}
