% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{generate_cicero_models}
\alias{generate_cicero_models}
\title{Generate \code{cicero} model}
\usage{
generate_cicero_models(
  cds,
  distance_parameter,
  s = 0.75,
  window = 5e+05,
  max_elements = 200,
  genomic_coords = NULL
)
}
\description{
Generate \code{cicero} model
}
\keyword{internal}
