% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{get_cts_en_GRNs}
\alias{get_cts_en_GRNs}
\title{Predict cell-type-specific enhancer-driven gene regulatory networks (eGRNs)}
\usage{
get_cts_en_GRNs(
  obj = NULL,
  celltype = NULL,
  en.regs = NULL,
  peak.assay = "ATAC",
  rna.dims = 50,
  atac.dims = 50,
  padj.cutoff = 0.05,
  out.dir = "./"
)
}
\arguments{
\item{obj}{An \code{Seurat} object used to get the eRegulons, NULL by default}

\item{celltype}{The metadata column indicating the cell types or clusters, NULL by default}

\item{en.regs}{A list of eRegulons, NULL by default}

\item{peak.assay}{The chromatin accessibility assay, "ATAC" by default}

\item{rna.dims}{The number of dimensions for RNA dimension reduction, 50 by default}

\item{atac.dims}{The number of dimensions for ATAC dimension reduction, 50 by default}

\item{padj.cutoff}{The cutoff of adjusted p-value of hyper-geometric test, 0.05 by default}

\item{out.dir}{The directory to save the intermediate results or final results, "./" by default}
}
\value{
Returns a list of eGRNs saved in \code{GRanges} object
}
\description{
Predict cell-type-specific enhancer-driven gene regulatory networks (eGRNs)
}
