% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{get_cts_en_regs}
\alias{get_cts_en_regs}
\title{Identify cell-type-specific enhancer regulons (eRegulons)}
\usage{
get_cts_en_regs(
  obj = NULL,
  peak.assay = "ATAC",
  de.genes = NULL,
  cts.en.grns = NULL,
  accessibility = FALSE,
  out.dir = "./",
  min.pct = 0.25,
  logfc.threshold = 0.25,
  padj.cutoff = 0.05
)
}
\arguments{
\item{obj}{An \code{Seurat} object used to get the eRegulons, NULL by default}

\item{peak.assay}{The chromatin accessibility assay, "ATAC" by default}

\item{de.genes}{A list of differentially expressed genes (DEGs), NULL by default}

\item{cts.en.grns}{Cell-type-specific eGRNs, NULL by default}

\item{accessibility}{whether perform differential accessibility analysis, FALSE by default}

\item{out.dir}{The directory to save the intermediate results or final results, "./" by default}

\item{padj.cutoff}{The cutoff of adjusted p-values of differential expression, 0.05 by default}

\item{celltype}{The metadata column indicating the cell types or clusters, NULL by default}
}
\value{
Returns a list of cell-type-specific eRegulons and plot dotplot-heatmap
}
\description{
Identify cell-type-specific enhancer regulons (eRegulons)
}
