% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{get_matrix_list}
\alias{get_matrix_list}
\title{Get the list of RNA and ATAC matrices}
\usage{
get_matrix_list(m, obj.list, assay = "RNA")
}
\description{
Get the list of RNA and ATAC matrices
}
\keyword{internal}
