% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{get_top_genes_peaks}
\alias{get_top_genes_peaks}
\title{Get the top-ranked genes, 100 by default}
\usage{
get_top_genes_peaks(
  HBC,
  cand.genes,
  cand.peaks,
  top.ngenes = 100,
  G,
  rna.m,
  atac.m
)
}
\description{
Get the top-ranked genes, 100 by default
}
\keyword{internal}
