% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{hybrid_biclust}
\alias{hybrid_biclust}
\title{Hybrid biclustering}
\usage{
hybrid_biclust(
  seeds = NULL,
  rna.list = NULL,
  atac.list = NULL,
  top.ngenes = 5,
  bound.TFs = NULL,
  same.terminal = F,
  binding.CREs = NULL,
  G.list = NULL,
  TFGene.pairs = NULL,
  c.cutoff = 1,
  KL = "min.exp",
  org.gs = org.gs,
  ego.order = 1,
  closure = F,
  intra.cutoff = 1,
  inter.cutoff = 0.8,
  peak.cutoff = 0.8,
  quantile.cutoff,
  rna.dis = NULL,
  atac.dis = NULL,
  min.cells = 10
)
}
\description{
Hybrid biclustering
}
\keyword{internal}
