% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{inter_eligible_seed}
\alias{inter_eligible_seed}
\title{Check whether a seed is eligible based on the similarity with other
seeds related to different TFs}
\usage{
inter_eligible_seed(s, h, gene.cutoff = 0.8, peak.cutoff = 0.8)
}
\description{
Check whether a seed is eligible based on the similarity with other
seeds related to different TFs
}
\keyword{internal}
