% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{intra_eligible_seed}
\alias{intra_eligible_seed}
\title{Whether the seed is eligible based on the similarity with other
seeds related to the same TF}
\usage{
intra_eligible_seed(s, h, gene.cutoff = 1)
}
\description{
Whether the seed is eligible based on the similarity with other
seeds related to the same TF
}
\keyword{internal}
