% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{load_annotation}
\alias{load_annotation}
\title{Load the annotation database of an organism}
\usage{
load_annotation(org = "hg38")
}
\description{
Load the annotation database of an organism
}
\keyword{internal}
