% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cistrome_tools.R
\name{load_database}
\alias{load_database}
\title{Load the genomic annotations of an organism}
\usage{
load_database(org = "hg38")
}
\description{
Load the genomic annotations of an organism
}
\keyword{internal}
