% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{make_cicero_cds}
\alias{make_cicero_cds}
\title{Make a \code{cicero} cell dataset}
\usage{
make_cicero_cds(
  cds,
  reduced_coordinates,
  k = 50,
  summary_stats = NULL,
  size_factor_normalize = TRUE,
  silent = FALSE,
  return_agg_info = FALSE
)
}
\description{
Make a \code{cicero} cell dataset
}
\keyword{internal}
