% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{MST_to_SFP}
\alias{MST_to_SFP}
\title{Construct Steiner forest based on minimum spanning trees}
\usage{
MST_to_SFP(G, ter)
}
\description{
Construct Steiner forest based on minimum spanning trees
}
\keyword{internal}
