% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{org_to_DB}
\alias{org_to_DB}
\title{Function org_to_DB used to get the database based on organism}
\usage{
org_to_DB(org = "hg38")
}
\description{
Function org_to_DB used to get the database based on organism
}
\keyword{internal}
