% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{overlap_peak_lst}
\alias{overlap_peak_lst}
\title{Calculate whether two list of \code{GRanges} are overlapped}
\usage{
overlap_peak_lst(lst1, lst2)
}
\description{
Calculate whether two list of \code{GRanges} are overlapped
}
\keyword{internal}
