% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{remove_steiner}
\alias{remove_steiner}
\title{Remove edges connecting two Steiner nodes, i.e., enhancers}
\usage{
remove_steiner(df)
}
\description{
Remove edges connecting two Steiner nodes, i.e., enhancers
}
\keyword{internal}
