% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{rna_atac_matrices_to_Seurat}
\alias{rna_atac_matrices_to_Seurat}
\title{Convert a pair of scRNA-seq and scATAC-seq mastrices into a \code{Seurat} object}
\usage{
rna_atac_matrices_to_Seurat(
  rna_counts = NULL,
  atac_counts = NULL,
  org = "hg38",
  frag.file = NULL,
  sep = c("-", "-")
)
}
\description{
Convert a pair of scRNA-seq and scATAC-seq mastrices into a \code{Seurat} object
}
\keyword{internal}
