% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{run_stream}
\alias{run_stream}
\title{Identify enhancer regulons (eRegulons) from jointly profiled scRNA-seq and scATAC-seq data}
\usage{
run_stream(
  obj = NULL,
  peak.assay = "ATAC",
  var.genes = 3000,
  top.peaks = 3000,
  min.cells = 10,
  out.dir = "./",
  org = "hg38",
  top.ngenes = 5,
  c.cutoff = 1,
  n.blocks = 100,
  distance = 250000,
  ifWeighted = TRUE,
  cicero.covar = -Inf,
  signac.score = -Inf,
  signac.pval = Inf,
  intra.cutoff = 1,
  inter.cutoff = 0.8,
  peak.cutoff = 0.8,
  score.cutoff = 1,
  KL = "min.exp",
  quantile.cutoff = 4,
  BlockOverlap = 0.5,
  Extension = 0.9,
  url.link = "https://figshare.com/ndownloader/files/38644505",
  submod.step = 30,
  min.eRegs = 100
)
}
\arguments{
\item{obj}{A \code{Seurat} object composed of both scRNA-sed and scATAC-seq assays}

\item{peak.assay}{The name of the scATAC-seq assay, "ATAC" by default}

\item{var.genes}{The number of highly variable genes to predict used to identify the core part of HBCs,
3000 by default}

\item{top.peaks}{The number of top-ranked peaks to identify the core part of hybrid biclusters (HBCs),
3000 by default}

\item{min.cells}{The cutoff of minimum number of cells for quality control (QC), 10 by default}

\item{out.dir}{The directory to save the intermediate or final results, "./" by default}

\item{org}{The organism version, hg38 by default}

\item{top.ngenes}{The number of genes composing the core part of an HBC, 5 by default}

\item{c.cutoff}{The cutoff of consistency during hybrid biclustering process, 1.0 by default}

\item{n.blocks}{The cutoff of the maximum number of blocks output by IRIS-FGM, 100 by default}

\item{distance}{The distance cutoff to build enhancer-enhancer relations, 250000 by default}

\item{intra.cutoff}{The cutoff to calculate pairwise similarity among HBCs associated with the same TFs,
1.0 by default}

\item{inter.cutoff}{The cutoff to compute pairwise similarity among genes in HBCs associated with different TFs,
0.80 by default}

\item{peak.cutoff}{The cutoff to quantify pairwise similarity among enhancers in HBCs, 0.80 by default}

\item{KL}{Which method to use for measuring the score of HBCs, "min.exp" by default, i.e.,
the smaller one between the numbers of genes and cells in a HBC}

\item{quantile.cutoff}{The quantile cutoff of the ratio of HBC cells, where enhancers are accessible,
4 by default, indicating the top-25\% among ranks}

\item{BlockOverlap}{The cutoff of maximum overlap between blocks output by \code{IRIS-FGM}, 0.500 by default}

\item{Extension}{The consistency level to expand a block by \code{IRIS-FGM}, 0.70 by default}

\item{submod.step}{The step size of the number of HBCs for submodular optimization, 30 by default}

\item{min.eRegs}{The minimum number of enhancer regulons (eRegulons) to output, 100 by default}
}
\value{
When running on a \code{Seurat} object,
returns a list of enhancer regulons, i.e., eRegulons, saved in a nested list
}
\description{
Identify enhancer regulons (eRegulons) from jointly profiled scRNA-seq and scATAC-seq data
}
\concept{run_stream}
