% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{score_HBC}
\alias{score_HBC}
\title{Score hybrid biclusters (HBCs)}
\usage{
score_HBC(
  HBC,
  m = NULL,
  KL = "min.exp",
  Q = NULL,
  P = NULL,
  G = NULL,
  VERY_SMALL = 1e-05
)
}
\description{
Score hybrid biclusters (HBCs)
}
\keyword{internal}
