% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{select_HBC}
\alias{select_HBC}
\title{Select the next hybrid bicluster (HBC) to add}
\usage{
select_HBC(HBCs, sim.m, HBC.flag, HBC.max.sim, step = 30, regulon.ids)
}
\description{
Select the next hybrid bicluster (HBC) to add
}
\keyword{internal}
