% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{sub_mod}
\alias{sub_mod}
\title{Perform submodular optimization}
\usage{
sub_mod(
  HBCs,
  sim.m,
  G.list,
  n.cells,
  rna.list,
  block.list,
  obj,
  peak.assay = "ATAC",
  links.df,
  min.eRegs = 100,
  step = 50
)
}
\description{
Perform submodular optimization
}
\keyword{internal}
