% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiome_tools.R
\name{subset_object}
\alias{subset_object}
\title{Subset a \code{Seurat} object}
\usage{
subset_object(
  LTMG.obj,
  object,
  peak.assay = "ATAC",
  atac.dis,
  max.peaks = 3000,
  links.df = NULL,
  n.blocks = 100
)
}
\description{
Subset a \code{Seurat} object
}
\keyword{internal}
