% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigenome_tools.R
\name{run_cicero}
\alias{run_cicero}
\title{Run Cicero}
\usage{
run_cicero(
  cds,
  genomic_coords,
  window = 5e+05,
  silent = FALSE,
  sample_num = 100
)
}
\description{
Run Cicero
}
\keyword{internal}
